Remove `ParsedContractClass`
