+++ src/core/contract_address/sierra_contract_address.rs
@@ -1,5 +1,5 @@
 use crate::core::errors::contract_address_errors::ContractAddressError;
-use crate::services::api::contract_classes::deprecated_parsed_contract_class::{
+use crate::services::api::contract_classes::deprecated_contract_class::{
     ContractEntryPoint, EntryPointType,
 };
 use cairo_lang_starknet::{
