+++ src/parser_errors.rs
@@ -1,4 +1,4 @@
-use crate::services::api::contract_classes::deprecated_parsed_contract_class::EntryPointType;
+use crate::services::api::contract_classes::deprecated_contract_class::EntryPointType;
 use crate::{
     core::errors::{contract_address_errors::ContractAddressError, state_errors::StateError},
     syscalls::syscall_handler_errors::SyscallHandlerError,
