+++ src/services/api/contract_classes/mod.rs
@@ -1,3 +1,2 @@
 pub mod compiled_class;
 pub mod deprecated_contract_class;
-pub(crate) mod deprecated_parsed_contract_class;
