+++ src/syscalls/business_logic_syscall_handler.rs
@@ -51,7 +51,7 @@ use cairo_vm::{
 };
 use lazy_static::lazy_static;
 
-use crate::services::api::contract_classes::deprecated_parsed_contract_class::EntryPointType;
+use crate::services::api::contract_classes::deprecated_contract_class::EntryPointType;
 use num_traits::{One, ToPrimitive, Zero};
 
 const STEP: u128 = 100;
