+++ src/transaction/invoke_function.rs
@@ -20,7 +20,7 @@ use crate::{
     utils::{calculate_tx_resources, Address},
 };
 
-use crate::services::api::contract_classes::deprecated_parsed_contract_class::EntryPointType;
+use crate::services::api::contract_classes::deprecated_contract_class::EntryPointType;
 use cairo_vm::felt::Felt252;
 use getset::Getters;
 use num_traits::Zero;
